Tests/LibWeb/Text/input/css/transition-basics.html

<!DOCTYPE html>
<html>
<script src="../include.js"></script>
<style>
    #box {
        width: 200px;
        height: 200px;
        background-color: red;
        position: absolute;
        left: 0;
        top: 0;

        transition: left 1s linear;
    }

    #box.move {
        left: 200px;
    }
</style>
<div id="box"></div>
<script>
    test(() => {
        let box = document.getElementById("box");
        println(`left starts at: ${getComputedStyle(box).left}`);

        box.classList.add("move");
        internals.updateAnimationsWithTimeOffset(1);

        let left = parseFloat(getComputedStyle(box).left);
        println(`shortly after starting, left is >= 0px? ${left >= 0} and < 200px? ${left < 200}`);

        internals.updateAnimationsWithTimeOffset(500);

        left = parseFloat(getComputedStyle(box).left);
        println(`half-way through, left is > 0px? ${left > 0} and < 200px? ${left < 200}`);

        internals.updateAnimationsWithTimeOffset(950);

        left = parseFloat(getComputedStyle(box).left);
        println(`near the end, left is > 0px? ${left > 0} and < 200px? ${left < 200}`);

        internals.updateAnimationsWithTimeOffset(1050);

        // FIXME: This is a hack to make it recompute the style, otherwise we get 0px for the left value.
        //        Figure out why!
        box.style.top = "0px";

        println(`after the transition, left is: ${getComputedStyle(box).left}`);
    });
</script>
</html>
